requirements.txt
google-generativeai>=0.3.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0